Fixes #1264 - CFN_GVAR_CST_MAX is 500, int16_t is needed

+++ companion/src/firmwares/opentx/opentxeeprom.cpp
@@ -1908,7 +1908,7 @@ class ArmCustomFunctionField: public TransformedField {
           else if (fn.adjustMode == 2)
             fn.param = RawSource(SOURCE_TYPE_GVAR, value).toValue();
           else
-            fn.param = (int8_t)value;
+            fn.param = (int16_t)value;
         }
         else if (version >= 214) {
           fn.adjustMode = _mode;
